created new checkpoints with materials
Created new Prefabs for the Checkpoints with 3D Buttons, since the 2D UI-Buttons wouldn't work in VR.
Added a trigger on the right hand, which detects the trigger with the button and enables them with the Script "Hand".
Assigned them the right color with new materials in the same colors as eastereggs, but without their texture.